+++ pycroscopy/io/__init__.py
@@ -1,5 +1,5 @@
-import hdfutils
-import ioutils
-import iohdf5
+import hdf_utils
+import io_utils
+import io_hdf5
 import microdata
 import translators
