+++ pycroscopy/io/translators/__init__.py
@@ -1,3 +1,3 @@
-import beutils
+import be_utils
 import utils
 import translator
