+++ pycroscopy/io/translators/fast_iv.py
@@ -13,8 +13,8 @@ import h5py
 from .translator import Translator
 from .utils import makePositionMat, getPositionSlicing, generateDummyMainParms
 from ..microdata import MicroDataGroup, MicroDataset # The building blocks for defining heirarchical storage in the H5 file
-from ..iohdf5 import ioHDF5 # Now the translator is responsible for writing the data.
-from ..hdfutils import getH5DsetRefs
+from ..io_hdf5 import ioHDF5 # Now the translator is responsible for writing the data.
+from ..hdf_utils import getH5DsetRefs
 
 class FastIVTranslator(Translator):
     """
